Extract list package
