+++ client/go/internal/list/array_list.go
@@ -1,8 +1,7 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 // Author: arnej
 
-// generic utilities
-package util
+package list
 
 type ArrayList[E comparable] []E
