+++ client/go/internal/list/array_list_test.go
@@ -1,5 +1,6 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package util
+
+package list
 
 import (
 	"testing"
